+++ src/test/java/org/opentripplanner/routing/algorithm/TurnCostTest.java
@@ -16,7 +16,8 @@ import org.opentripplanner.routing.core.RoutingContext;
 import org.opentripplanner.routing.core.State;
 import org.opentripplanner.routing.core.TraverseMode;
 import org.opentripplanner.routing.core.TraverseModeSet;
-import org.opentripplanner.routing.core.intersection_model.ConstantIntersectionTraversalCostModel;
+import org.opentripplanner.routing.core.intersection_model.ConstantIntersectionTraversalCalculator;
+import org.opentripplanner.routing.core.intersection_model.IntersectionTraversalCalculator;
 import org.opentripplanner.routing.edgetype.StreetEdge;
 import org.opentripplanner.routing.edgetype.StreetTraversalPermission;
 import org.opentripplanner.routing.graph.Graph;
@@ -38,6 +39,8 @@ public class TurnCostTest {
   private StreetEdge maple_main1, broad1_2;
 
   private RouteRequest proto;
+
+  private IntersectionTraversalCalculator calculator;
 
   @BeforeEach
   public void before() {
@@ -101,7 +104,7 @@ public class TurnCostTest {
     preferences.setAllStreetReluctance(1.0);
 
     // Turn costs are all 0 by default.
-    graph.setIntersectionTraversalCostModel(new ConstantIntersectionTraversalCostModel(0.0));
+    calculator = new ConstantIntersectionTraversalCalculator(0.0);
   }
 
   @Test
@@ -112,13 +115,12 @@ public class TurnCostTest {
 
   @Test
   public void testForwardDefaultConstTurnCosts() {
-    RouteRequest options = proto.clone();
-    graph.setIntersectionTraversalCostModel(new ConstantIntersectionTraversalCostModel(10.0));
+    calculator = new ConstantIntersectionTraversalCalculator(10.0);
 
     // Without turn costs, this path costs 2x100 + 2x50 = 300.
     // Since we traverse 3 intersections, the total cost should be 330.
     GraphPath path = checkForwardRouteDuration(
-      new RoutingContext(options, graph, topRight, bottomLeft),
+      new RoutingContext(proto, graph, topRight, bottomLeft),
       330
     );
 
@@ -164,7 +166,7 @@ public class TurnCostTest {
   @Test
   public void testForwardCarConstTurnCosts() {
     RouteRequest options = proto.clone();
-    graph.setIntersectionTraversalCostModel(new ConstantIntersectionTraversalCostModel(10.0));
+    calculator = new ConstantIntersectionTraversalCalculator(10.0);
     options.setMode(TraverseMode.CAR);
 
     // Without turn costs, this path costs 3x100 + 1x50 = 350.
@@ -191,7 +193,11 @@ public class TurnCostTest {
   }
 
   private GraphPath checkForwardRouteDuration(RoutingContext context, int expectedDuration) {
-    ShortestPathTree tree = AStarBuilder.oneToOne().setContext(context).getShortestPathTree();
+    ShortestPathTree tree = AStarBuilder
+      .oneToOne()
+      .setContext(context)
+      .setIntersectionTraversalCalculator(calculator)
+      .getShortestPathTree();
     GraphPath path = tree.getPath(bottomLeft);
     assertNotNull(path);