+++ docs/Changelog.md
@@ -106,6 +106,7 @@ based on merged pull requests. Search GitHub issues and pull requests for smalle
 - Add mapping for RecordedCall.PredictionInaccurate [#4435](https://github.com/opentripplanner/OpenTripPlanner/pull/4435)
 - Routing request refactor - routing preferences [#4426](https://github.com/opentripplanner/OpenTripPlanner/pull/4426)
 - Extract journey request from RouteRequest [#4446](https://github.com/opentripplanner/OpenTripPlanner/pull/4446)
+- Move intersectionTraversalCostModel into StreetPreferences [#4444](https://github.com/opentripplanner/OpenTripPlanner/pull/4444)
   [](AUTOMATIC_CHANGELOG_PLACEHOLDER_DO_NOT_REMOVE)
 
 ## 2.1.0 (2022-03-17)